// front-end/dashboard/src/api-client/models/FinancialRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Management Service
 * API for managing tasks for patients, doctors, and administrators.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FinancialRecord
 */
export interface FinancialRecord {
    /**
     * 
     * @type {string}
     * @memberof FinancialRecord
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof FinancialRecord
     */
    consultationType?: string;
    /**
     * 
     * @type {number}
     * @memberof FinancialRecord
     */
    patientCount?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialRecord
     */
    taskCount?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialRecord
     */
    totalRevenue?: number;
    /**
     * 
     * @type {string}
     * @memberof FinancialRecord
     */
    duration?: string;
    /**
     * 
     * @type {string}
     * @memberof FinancialRecord
     */
    updatedAt?: string;
}

/**
 * Check if a given object implements the FinancialRecord interface.
 */
export function instanceOfFinancialRecord(value: object): value is FinancialRecord {
    return true;
}

export function FinancialRecordFromJSON(json: any): FinancialRecord {
    return FinancialRecordFromJSONTyped(json, false);
}

export function FinancialRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): FinancialRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'consultationType': json['consultationType'] == null ? undefined : json['consultationType'],
        'patientCount': json['patientCount'] == null ? undefined : json['patientCount'],
        'taskCount': json['taskCount'] == null ? undefined : json['taskCount'],
        'totalRevenue': json['totalRevenue'] == null ? undefined : json['totalRevenue'],
        'duration': json['duration'] == null ? undefined : json['duration'],
        'updatedAt': json['updatedAt'] == null ? undefined : json['updatedAt'],
    };
}

export function FinancialRecordToJSON(json: any): FinancialRecord {
    return FinancialRecordToJSONTyped(json, false);
}

export function FinancialRecordToJSONTyped(value?: FinancialRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'consultationType': value['consultationType'],
        'patientCount': value['patientCount'],
        'taskCount': value['taskCount'],
        'totalRevenue': value['totalRevenue'],
        'duration': value['duration'],
        'updatedAt': value['updatedAt'],
    };
}